requirements.txt
fastapi>=0.115.0,<1
uvicorn[standard]>=0.32.0,<1
python-dotenv>=1.0.1,<2
